Create Bottom-Up Approach
This is an example of the bottom-up approach to multilevel modeling using R.  In this situation, we start with either theoretically-relevant level 1 variables or exploratory data analysis.  We follow this up with an examination of random and fixed effects.  After finalizing this model with the level 1 variables, we will finish up by using level 2 variables.